+++ src/main/webapp/precache-manifest.483c083b84bcb465cfc2548676d6497f.js
@@ -1,0 +1,34 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "515ec0a809e96de650b2addbec5a1e7c",
+    "url": "/index.html"
+  },
+  {
+    "revision": "34ab0f3e335d3a9f91da",
+    "url": "/static/css/main.2ad0accb.chunk.css"
+  },
+  {
+    "revision": "b8578e9fe29de027b597",
+    "url": "/static/js/2.5c7fbf52.chunk.js"
+  },
+  {
+    "revision": "89b0379e7bcda1a468d8b0343aeb4e53",
+    "url": "/static/js/2.5c7fbf52.chunk.js.LICENSE.txt"
+  },
+  {
+    "revision": "34ab0f3e335d3a9f91da",
+    "url": "/static/js/main.32f660cd.chunk.js"
+  },
+  {
+    "revision": "77a788e63b3c02a41b00",
+    "url": "/static/js/runtime-main.2ecb3fdc.js"
+  },
+  {
+    "revision": "5db9fea0ec9e05cfb98e7387be5d0aa7",
+    "url": "/static/media/gb.5db9fea0.svg"
+  },
+  {
+    "revision": "8ec583188aba7e9426580350312d97a5",
+    "url": "/static/media/us.8ec58318.svg"
+  }
+]);