+++ src/main/webapp/precache-manifest.e24900b1f3edc4f53d4013a24ba49980.js
@@ -1,0 +1,34 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "f00e3621fb3ab3f2ec9094854e3338e5",
+    "url": "/index.html"
+  },
+  {
+    "revision": "b17e39383e7449a327bf",
+    "url": "/static/css/main.2ad0accb.chunk.css"
+  },
+  {
+    "revision": "b8578e9fe29de027b597",
+    "url": "/static/js/2.5c7fbf52.chunk.js"
+  },
+  {
+    "revision": "89b0379e7bcda1a468d8b0343aeb4e53",
+    "url": "/static/js/2.5c7fbf52.chunk.js.LICENSE.txt"
+  },
+  {
+    "revision": "b17e39383e7449a327bf",
+    "url": "/static/js/main.a3f80faa.chunk.js"
+  },
+  {
+    "revision": "77a788e63b3c02a41b00",
+    "url": "/static/js/runtime-main.2ecb3fdc.js"
+  },
+  {
+    "revision": "5db9fea0ec9e05cfb98e7387be5d0aa7",
+    "url": "/static/media/gb.5db9fea0.svg"
+  },
+  {
+    "revision": "8ec583188aba7e9426580350312d97a5",
+    "url": "/static/media/us.8ec58318.svg"
+  }
+]);